apply feedback from review
