+++ config/emails.yaml
@@ -17,3 +17,11 @@ services:
 
   Pimcore\Bundle\StudioBackendBundle\Email\Repository\BlocklistRepositoryInterface:
     class: Pimcore\Bundle\StudioBackendBundle\Email\Repository\BlocklistRepository
+
+
+  #
+  # Services
+  #
+
+  Pimcore\Bundle\StudioBackendBundle\Email\Service\BlocklistServiceInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Email\Service\BlocklistService
