+++ src/Email/Controller/Blocklist/AddController.php
@@ -19,8 +19,8 @@ namespace Pimcore\Bundle\StudioBackendBundle\Email\Controller\Blocklist;
 use OpenApi\Attributes\Post;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\Email\Attributes\Request\BlocklistRequestBody;
-use Pimcore\Bundle\StudioBackendBundle\Email\Repository\BlocklistRepositoryInterface;
 use Pimcore\Bundle\StudioBackendBundle\Email\Schema\BlocklistEntryRequest;
+use Pimcore\Bundle\StudioBackendBundle\Email\Service\BlocklistServiceInterface;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\AccessDeniedException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\EnvironmentException;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attributes\Response\DefaultResponses;
@@ -41,7 +41,7 @@ final class AddController extends AbstractApiController
 {
     public function __construct(
         SerializerInterface $serializer,
-        private readonly BlocklistRepositoryInterface $blocklistRepository,
+        private readonly BlocklistServiceInterface $blocklistService,
     ) {
         parent::__construct($serializer);
     }
@@ -68,7 +68,7 @@ final class AddController extends AbstractApiController
     public function addBlocklistEntry(
         #[MapRequestPayload] BlocklistEntryRequest $request
     ): Response {
-        $this->blocklistRepository->addEntry($request->getEmail());
+        $this->blocklistService->addEntry($request->getEmail());
 
         return new Response();
     }
