+++ src/OpenApi/Config/Tags.php
@@ -46,7 +46,7 @@ use OpenApi\Attributes\Tag;
     description: 'Get element properties for a single element based on its type and provided parameters.'
 )]
 #[Tag(
-    name: Tags::Emails->name,
+    name: Tags::Emails->value,
     description: 'Email operations to get/update/create/delete/test emails and email blocklist.'
 )]
 #[Tag(
@@ -101,7 +101,7 @@ enum Tags: string
     case DataObjects = 'DataObjects';
     case Dependencies = 'Dependencies';
     case Elements = 'Elements';
-    case Emails = 'Emails';
+    case Emails = 'E-Mails';
     case Grid = 'Grid';
     case Mercure = 'Mercure';
     case Notes = 'Notes';
